+++ src/app/AttributeReportBuilder.cpp
@@ -1,6 +1,5 @@
 /*
- *
- *    Copyright (c) 2021-2022 Project CHIP Authors
+ *    Copyright (c) 2024 Project CHIP Authors
  *    All rights reserved.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
