+++ src/app/AttributeValueEncoder.cpp
@@ -1,6 +1,5 @@
 /*
- *
- *    Copyright (c) 2021 Project CHIP Authors
+ *    Copyright (c) 2024 Project CHIP Authors
  *    All rights reserved.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,45 +14,10 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-
-#include <app/AttributeAccessInterface.h>
+#include <app/AttributeValueEncoder.h>
 
 namespace chip {
 namespace app {
-
-CHIP_ERROR AttributeReportBuilder::PrepareAttribute(AttributeReportIBs::Builder & aAttributeReportIBsBuilder,
-                                                    const ConcreteDataAttributePath & aPath, DataVersion aDataVersion)
-{
-    AttributeReportIB::Builder & attributeReportIBBuilder = aAttributeReportIBsBuilder.CreateAttributeReport();
-    ReturnErrorOnFailure(aAttributeReportIBsBuilder.GetError());
-
-    AttributeDataIB::Builder & attributeDataIBBuilder = attributeReportIBBuilder.CreateAttributeData();
-    ReturnErrorOnFailure(attributeReportIBBuilder.GetError());
-
-    attributeDataIBBuilder.DataVersion(aDataVersion);
-
-    AttributePathIB::Builder & attributePathIBBuilder = attributeDataIBBuilder.CreatePath();
-    ReturnErrorOnFailure(attributeDataIBBuilder.GetError());
-
-    attributePathIBBuilder.Endpoint(aPath.mEndpointId).Cluster(aPath.mClusterId).Attribute(aPath.mAttributeId);
-
-    if (aPath.mListOp == ConcreteDataAttributePath::ListOperation::AppendItem)
-    {
-        // An append to a list (or a data chunk consisting just one list entry that's part of a bigger list) is represented by a
-        // null list index in the path.
-        attributePathIBBuilder.ListIndex(DataModel::Nullable<ListIndex>());
-    }
-
-    ReturnErrorOnFailure(attributePathIBBuilder.EndOfAttributePathIB());
-
-    return attributeDataIBBuilder.GetError();
-}
-
-CHIP_ERROR AttributeReportBuilder::FinishAttribute(AttributeReportIBs::Builder & aAttributeReportIBsBuilder)
-{
-    ReturnErrorOnFailure(aAttributeReportIBsBuilder.GetAttributeReport().GetAttributeData().EndOfAttributeDataIB());
-    return aAttributeReportIBsBuilder.GetAttributeReport().EndOfAttributeReportIB();
-}
 
 namespace {
